Ver 0.7.9
- More fixes for Linux-Lite.

+++ Contrib/at67/audio.cpp
@@ -138,7 +138,7 @@ namespace Audio
                 {
                     uint16_t segment = *_scorePtr++;
                     segment |= (((*_scorePtr++) <<8) & 0xFF00);
-                    _scorePtr = (uint8_t*)segment;
+                    _scorePtr = (uint8_t*)(uintptr_t)segment;
                 }
             }
             // Delay n milliseconds where n = 8bit value
